backend/requirements.txt
setuptools>=65.5.1
undetected-chromedriver==3.5.5
selenium==4.18.1
webdriver-manager==4.0.1
annotated-types==0.7.0
anyio==4.6.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
click==8.1.7
dnspython==2.7.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
pymongo[srv]==4.10.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
cloudflare==2.8.15
httpx==0.24.1
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
huggingface-hub==0.20.3
Pillow==10.2.0
Werkzeug==3.0.1
python-jwt==4.1.0
